+++ app/src/main/java/com/taksapp/taksapp/application/arch/utils/Result.kt
@@ -1,4 +1,4 @@
-package com.taksapp.taksapp.arch.utils
+package com.taksapp.taksapp.application.arch.utils
 
 enum class Status {
     LOADING,
