+++ app/src/main/java/com/taksapp/taksapp/application/auth/viewmodels/LoginViewModel.kt
@@ -1,4 +1,4 @@
-package com.taksapp.taksapp.ui.auth.viewmodels
+package com.taksapp.taksapp.application.auth.viewmodels
 
 import android.content.Context
 import androidx.lifecycle.LiveData
@@ -6,8 +6,8 @@ import androidx.lifecycle.MutableLiveData
 import androidx.lifecycle.Observer
 import androidx.lifecycle.ViewModel
 import com.taksapp.taksapp.R
-import com.taksapp.taksapp.arch.utils.Event
-import com.taksapp.taksapp.arch.utils.Result
+import com.taksapp.taksapp.application.arch.utils.Event
+import com.taksapp.taksapp.application.arch.utils.Result
 import com.taksapp.taksapp.data.repositories.AuthenticationRepository
 import com.taksapp.taksapp.data.repositories.LoginError
