+++ app/src/main/java/com/taksapp/taksapp/application/taxirequest/presentationmodels/LocationPresentationModel.kt
@@ -1,3 +1,3 @@
-package com.taksapp.taksapp.ui.taxi.presentationmodels
+package com.taksapp.taksapp.application.taxirequest.presentationmodels
 
 data class LocationPresentationModel(val latitude: Double, val longitude: Double)
