+++ app/src/main/java/com/taksapp/taksapp/application/taxirequest/presentationmodels/PlacePresentationModel.kt
@@ -1,4 +1,4 @@
-package com.taksapp.taksapp.ui.taxi.presentationmodels
+package com.taksapp.taksapp.application.taxirequest.presentationmodels
 
 import androidx.recyclerview.widget.DiffUtil
 import java.io.Serializable
