+++ app/src/main/java/com/taksapp/taksapp/data/webservices/RidersTaxiRequestWebService.kt
@@ -1,6 +1,6 @@
 package com.taksapp.taksapp.data.webservices
 
-import com.taksapp.taksapp.arch.utils.Result
+import com.taksapp.taksapp.application.arch.utils.Result
 import com.taksapp.taksapp.data.webservices.client.Taksapp
 import com.taksapp.taksapp.data.webservices.client.resources.riders.*
 import com.taksapp.taksapp.domain.Location
