+++ app/src/main/java/com/taksapp/taksapp/di/modules/InfrastructureServicesModule.kt
@@ -1,7 +1,7 @@
 package com.taksapp.taksapp.di.modules
 
-import com.taksapp.taksapp.data.infrastructure.FirebasePushNotificationTokenRetriever
-import com.taksapp.taksapp.data.infrastructure.PushNotificationTokenRetriever
+import com.taksapp.taksapp.data.infrastructure.services.FirebasePushNotificationTokenRetriever
+import com.taksapp.taksapp.data.infrastructure.services.PushNotificationTokenRetriever
 import org.koin.dsl.module
 
 val infrastructureServicesModule = module {
