+++ app/src/main/java/com/taksapp/taksapp/di/modules/ViewModelsModule.kt
@@ -1,13 +1,13 @@
 package com.taksapp.taksapp.di.modules
 
 import com.taksapp.taksapp.domain.TaxiRequest
-import com.taksapp.taksapp.ui.auth.viewmodels.LoginViewModel
-import com.taksapp.taksapp.ui.auth.viewmodels.RiderSignUpOtpConfirmationViewModel
-import com.taksapp.taksapp.ui.auth.viewmodels.RiderSignUpViewModel
-import com.taksapp.taksapp.ui.launch.viewmodels.LaunchViewModel
-import com.taksapp.taksapp.ui.taxi.viewmodels.AutocompletePlaceChooserViewModel
-import com.taksapp.taksapp.ui.taxi.viewmodels.FareEstimationViewModel
-import com.taksapp.taksapp.ui.taxi.viewmodels.TaxiRequestViewModel
+import com.taksapp.taksapp.application.auth.viewmodels.LoginViewModel
+import com.taksapp.taksapp.application.auth.viewmodels.RiderSignUpOtpConfirmationViewModel
+import com.taksapp.taksapp.application.auth.viewmodels.RiderSignUpViewModel
+import com.taksapp.taksapp.application.launch.viewmodels.LaunchViewModel
+import com.taksapp.taksapp.application.taxirequest.viewmodels.AutocompletePlaceChooserViewModel
+import com.taksapp.taksapp.application.taxirequest.viewmodels.FareEstimationViewModel
+import com.taksapp.taksapp.application.taxirequest.viewmodels.TaxiRequestViewModel
 import org.koin.android.ext.koin.androidContext
 import org.koin.android.viewmodel.dsl.viewModel
 import org.koin.dsl.module
