+++ app/src/main/java/com/taksapp/taksapp/domain/interfaces/FareRepository.kt
@@ -1,7 +1,7 @@
 package com.taksapp.taksapp.domain.interfaces
 
 import androidx.lifecycle.LiveData
-import com.taksapp.taksapp.arch.utils.Result
+import com.taksapp.taksapp.application.arch.utils.Result
 import com.taksapp.taksapp.domain.FareEstimation
 import com.taksapp.taksapp.domain.Location
