+++ app/src/main/java/com/taksapp/taksapp/ui/auth/DriverLoginActivity.kt
@@ -9,7 +9,7 @@ import androidx.lifecycle.Observer
 import com.google.android.material.snackbar.Snackbar
 import com.taksapp.taksapp.R
 import com.taksapp.taksapp.databinding.ActivityDriverLoginBinding
-import com.taksapp.taksapp.ui.auth.viewmodels.LoginViewModel
+import com.taksapp.taksapp.application.auth.viewmodels.LoginViewModel
 import org.koin.android.viewmodel.ext.android.viewModel
 
 class DriverLoginActivity : AppCompatActivity() {
