+++ app/src/main/res/layout/activity_rider_sign_up_otp_confirmation.xml
@@ -7,7 +7,7 @@
         <import type="android.view.View" />
         <variable
             name="viewModel"
-            type="com.taksapp.taksapp.ui.auth.viewmodels.RiderSignUpOtpConfirmationViewModel" />
+            type="com.taksapp.taksapp.application.auth.viewmodels.RiderSignUpOtpConfirmationViewModel" />
     </data>
     
     <androidx.constraintlayout.widget.ConstraintLayout
