+++ app/src/main/res/layout/fragment_fares_estimates.xml
@@ -7,7 +7,7 @@
         <import type="android.view.View" />
         <variable
             name="viewModel"
-            type="com.taksapp.taksapp.ui.taxi.viewmodels.FareEstimationViewModel" />
+            type="com.taksapp.taksapp.application.taxirequest.viewmodels.FareEstimationViewModel" />
     </data>
     
     <androidx.cardview.widget.CardView
