+++ app/src/main/res/layout/fragment_locations_selection.xml
@@ -6,7 +6,7 @@
     
     <data>
         <import type="android.view.View" />
-        <variable name="viewModel" type="com.taksapp.taksapp.ui.taxi.viewmodels.FareEstimationViewModel" />
+        <variable name="viewModel" type="com.taksapp.taksapp.application.taxirequest.viewmodels.FareEstimationViewModel" />
     </data>
     
     <androidx.cardview.widget.CardView
