+++ app/src/test/java/com/taksapp/taksapp/application/taxirequest/viewmodels/FareEstimationViewModelTests.kt
@@ -1,11 +1,11 @@
-package com.taksapp.taksapp.ui.taxi.viewmodels
+package com.taksapp.taksapp.application.taxirequest.viewmodels
 
 import android.content.Context
 import androidx.arch.core.executor.testing.InstantTaskExecutorRule
 import com.nhaarman.mockitokotlin2.*
 import com.taksapp.taksapp.R
-import com.taksapp.taksapp.arch.utils.Result
-import com.taksapp.taksapp.data.infrastructure.PushNotificationTokenRetriever
+import com.taksapp.taksapp.application.arch.utils.Result
+import com.taksapp.taksapp.data.infrastructure.services.PushNotificationTokenRetriever
 import com.taksapp.taksapp.data.repositories.RiderTaxiRequestsRepository
 import com.taksapp.taksapp.domain.Location
 import com.taksapp.taksapp.domain.Status
@@ -14,7 +14,7 @@ import com.taksapp.taksapp.domain.interfaces.DevicesService
 import com.taksapp.taksapp.domain.interfaces.FareRepository
 import com.taksapp.taksapp.domain.interfaces.RidersTaxiRequestService
 import com.taksapp.taksapp.domain.interfaces.TaxiRequestError
-import com.taksapp.taksapp.ui.taxi.presentationmodels.PlacePresentationModel
+import com.taksapp.taksapp.application.taxirequest.presentationmodels.PlacePresentationModel
 import com.taksapp.taksapp.utils.MainCoroutineScopeRule
 import com.taksapp.taksapp.utils.getOrAwaitValue
 import junitparams.JUnitParamsRunner
